backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
transformers = "^4.27.4"
torch = "^2.0.0"
ipython = "^8.12.0"
Flask-Pydantic = "^0.11.0"
beautifulsoup4 = "^4.12.0"
pandas = "^1.5.3"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
